Provide ability to use custom annotations
This change simplify extending default Page Object implementation
with custom annotations. AbstractAnnotations class provides basic
methods for default annotations such as FindBy. It's extension can
be passed now to existing element locators, so there is no need to
implement them as well.
